Moved script to `src` folder
